MB-24765: DefragmenterVisitor: Split pause/resume handling into seperate class
Split the code to handle pausing in the middle of a HashTable and
resuming from (approximately) that point from DefragmenterVisitor into
a seperate class PauseResumeVBAdapter.

This allows other visitors (e.g. Ephemeral VBTombstonePurger) to make
use of this functionality.

As part of this, rename PauseResumeEPStoreVisitor ->
PauseResumeVBVisitor as that more accurately reflects what it does.

Change-Id: I3b5388b83dce37b00cb948b6d2aa7b5d343e8d3e
Reviewed-on: http://review.couchbase.org/79314
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>