MB-16181: Only scan the key if it's tagged as Collections
Move the key scanning function to be in inside the collections
branch, thus only users of collections will see the cost of scanning
for collections.

This was observed whilst profiling a single worker thread under heavy
write load.

doesKeyContainValidCollection hitting ~4% now the same workload it
doesn't show up in the profile report.

Change-Id: I092f188dee82822e32f60440b91693b480eb094a
Reviewed-on: http://review.couchbase.org/79186
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Will Gardner <willg@rdner.io>
Tested-by: Build Bot <build@couchbase.com>