Add support for dropping a privilege from the current context
During unit tests we need to test a command with and without the
appropriate privilege available in the current context. Being able
to drop a privilege from the current execution context makes
writing unit tests a lot easier as the alternative is to either
operate on two different buckets (or reauthenticate as a user
without the privilege).

Given that dropping a privilege from the execution context won't
increase a clients access, it is a safe operation to allow clients
to use.

The privilege set is reset if:

   * Someone changes the RBAC database
   * The client selects another bucket

Change-Id: I36d29cfc20dc3912b4a128cb5901902531a315dd
Reviewed-on: http://review.couchbase.org/79189
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>