MB-23529: Fix doWorkerStat data race
This race appears to be caused by GGC's buggy string COW as seen
multiple times, e.g., MB-23454.

doWorkerStat calls getLog/getSlowLog to get a vector of TaskLogEntrys,
which have been copied out of the tasklog ringbuffer of a given
ExecutorThread. These copies logically have copies of the original's
`std::string name`.

Once the ringbuffer is filled and begins overwriting older entries, the
deletion of the old `std::string name` races with doWorkerStats reading
the COW'd copied name.

This is resolved in this patch by adding an explicit copy constructor to
TaskLogEntry which forces a copy of the std::string name (in the same
manner as MB-23454).

Change-Id: Ib5d12fba7dced94b1792e12edd4c3a85a67f3362
Reviewed-on: http://review.couchbase.org/79194
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>