Refactor: CacheCallback to not directly access HashTable object
The hash table object should ideally be a protected member of the
VBucket class, as opposed to being public.

The patch moves towards this goal by introducing a getItem function to
the VBucket class, which ensures that the backfill function
CacheCallback::callback no longer requires direct access to the hash
table object.

Tests to ensure the refactor does not change functionality were added
in the following patch http://review.couchbase.org/#/c/78650/

Change-Id: I2c71dd30f6b8c66de668137cac1bf358bbdf6588
Reviewed-on: http://review.couchbase.org/79254
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>