MB-24631: Do not block on backfill worker thread
We have finite number of worker threads in our system. DCP backfill
is scheduled on worker threads. In case of multiple backfills on
the same vbucket we allow only one at a time. If the other backfills
do a blocking wait (1) number of worker threads doing actual work
is reduced, (2) we could run into a deadlock if the running backfill
on vbucket is paused (while holding the lock) and another backfill
is run by the thread which would then block on the same lock.

This commit, addresses the problem by doing non-blocking waits if
we are to run multiple backfills on the same vbucket. This is
achieved by returning optional range iterators when a request
to read the vbucket seq list is made. Upon the condition that there
is already one ongoing read on the sequence list, null iterator
is returned and the backfill module defers the backfill to a
later time.

Change-Id: I53108ec63bf443d18cd16a04a97d32f68fdb3aec
Reviewed-on: http://review.couchbase.org/79273
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>