Resolve couchstore compiler warning
couchstore-gerrit-master CV when running under Ubuntu 16.04 (clang-3.9)
warns regarding an defined but unused variable.

This variable is from gtest and is defined as part of
REGISTER_TYPED_TEST_CASE_P. BufferedWrappedOpsTest is registered in
couchstore/tests/wrapped_fileops_test.h and is used in couchstore tests,
but is also included in a kv_engine test, couch-fs-stats-test.cc.

BufferedWrappedOpsTest is instantiated elsewhere, but is not needed in
the kv_engine test. This patch instantiates it with an empty type list,
preventing the warning.

Change-Id: I00244b4fc0a19d97a2ff95674908565a61088d00
Reviewed-on: http://review.couchbase.org/79474
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>