MB-24765: Make EphTombstoneHTCleaner visit HashTable incrementally
Change the EphTombstoneHTCleaner task to visit the HashTable
incrementally, yielding back to the scheduler every
ephemeral_metadata_purge_chunk_duration (default: 20ms). This
significantly reduces the impact this background task may have on the
scheduling of other NonIO tasks, given that prevously it was seen to
take a number of seconds to run.

This is implemented in a similar way to the DefragmenterTask - use
HashTable::pauseResumeVisit() instead of ::visit() to perform the
visitation of each HashTable (which allows the visitor to yield and
pause execution).

One difference with the DefragmenterTask is that EphTombstoneHTCleaner
does *not* snooze between runs if it hasn't yet completed - it simply
returns true, which will cause the scheduler to re-schedule it as soon
as possible. The reason for this is that tombstone purging can make a
material difference on the size of data stored - and hence we want to
complete a visitation of the whole bucket as quickly as possible.

Change-Id: If07eb6a0845dffa6e49670731fab01479f266c06
Reviewed-on: http://review.couchbase.org/79555
Reviewed-by: Manu Dhundi <manu@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>