MB-22163: Recover value-eviction performance loss with store_if
Introduce a new engine method called store_if, this method
is exactly like store but takes a predicate (std::function).

For set and replace, they will call the predicate on the existing
item (if any) and if the predicate returns true, the store will
continue.

Memcached is changed to make use of store_if so that it can avoid
a get for determining if an existing item has xattrs.

store_if is tried first and if it fails because the existing item
has xattrs, then the slower read/modify/write path is chosen.

This recovers approx 10% in op/s primarily because we reduce the
hash-table locking (your mileage may vary).

Change-Id: I90af507695e4ae5afa195259cf4b7612f87de365
Reviewed-on: http://review.couchbase.org/79490
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>