MB-24817: During takeover, hold stream lock until vb is set to dead
During DCP takeover, when we decide to set the vbucket to dead state
on the producer side, we also transition the stream state from
STREAM_TAKEOVER_WAIT to STREAM_TAKEOVER_SEND state to send out any
remaining items that were received before the vbucket was set to
dead state.

The stream lock must be held until the vbucket is set to the dead
state, so that we do not prematurely finish sending out the last
items when the vbucket is not dead yet (that is if the vbucket
is not dead, it could get some more items).

This commit addresses the issue and also handles the ordering of
locks involved at stream level, ep store level and vbucket level
in the scenario.

Change-Id: I89bb42edec4f3765c8a9c67e6e89e9680eb40875
Reviewed-on: http://review.couchbase.org/79532
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>