Fixup error message when failing to set max filehandles
The string is multiline (and hard to read in a source editor), but
it turns out that there was cases where one line ended with a space,
and the next line started with a space resulting in two spaces between
the words.

This patch makes sure that all lines _end_ with a space, and the next
line does not start with a space.

Change-Id: I9e7301d821c827c945339ad7108289d0b8f986c8
Reviewed-on: http://review.couchbase.org/79592
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>