MB-24874: Partial revert of "Add GlobalTask::wakeUp, remove unnecessary snooze() calls"
This partially reverts commit
e158d66e9e927ba6ebdfed83d438c464bef3eba7.

The changes to move calls to snooze() have triggered a potential hang
on the DCP Processor task, due to a race between Processor::run()
snoozing forever, and an external actor calling notifyVbucketReady()
to wakeup the task. The effect of this race is that the wakeup() can
be missed, causing the Processor task to miss a notification and never
wakeup.

Revert the changes to snooze() ordering; keeps the new wakeUp method.

Change-Id: Ibb5109911fd35e2eb370d1a3720ab9a0898356b4
Reviewed-on: http://review.couchbase.org/79608
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>