MB-24363 Add ewouldblock functionality to trace dump chunk generation
Prior to this patch trace dump chunks were generated on a front-end
thread which would potentially block the thread while a multi-megabyte
string was being generated.

This patch switches to the chunks being generated on an executor
thread instead and adds ewouldblock handling to ioctl in general
to allow this to happen.

Change-Id: I2fa07e562b87d93fe69ec40a422e81fd83e5bc82
Reviewed-on: http://review.couchbase.org/79619
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>