MB-24889: Send datatype==raw_bytes for DCP_DELETION
For deletions of documents which were previously non-binary
(e.g. JSON), a datatype which wasn't RAW_BYTES could be sent via a
DCP_DELETION message. This would cause the replication consumer to
reject the DCP_DELETION:

    Invalid format specified for DCP_DELETION - 4 - closing connection

This would cause replication to fail.

This issue is triggered by a combination of factors:

1. Prior to 54ce79e, DCP_DELETION datatype field was handled by
   memcached, and was always RAW_BYTES[*].

2. Prior to 302b83e, the datatype was always held in the Blob of the
   StoredValue - as such if the value was reset (as happens during a
   delete) then the datatype read as RAW_BYTES. However this patch
   failed to clear the datatype (back to RAW_BYTES) when the value was
   reset.

Either one of the two changes doesn't cause the issue, but with both
the datatype is incorrectly left sent out over the wire as whatever it
was prior to deletion.

Fix by making StoredValue::reset() clear the datatype. Update an XDCR
test which was incorrectly assuming the datatype was not changed by
delete.

[*]: Note if a DCP_DELETION has XATTRs then it is permitted to have a
     datatype of XATTR. There is still an outstanding issue to address
     - the consumer's DCP_DELETION validator will reject any
     non-RAW_BYTES datatype.

Change-Id: Icff8d3afb3dd984db457895394dd2447b259582a
Reviewed-on: http://review.couchbase.org/79668
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Will Gardner <willg@rdner.io>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>