MB-24818: permit add_with_meta on deleted item in ephemeral
test_add_meta_conflict_resolution failed on an ephemeral bucket as
VBucket::processSet incorrectly reported MutationStatus::InvalidCas.

In a persistent bucket, when performing an add_with_meta against a
deleted item, the "old" StoredValue from the hash table is a temporary
deleted item. We consider temporary items (isTempItem) to "not exist";
the add operation is permitted.

In ephemeral, there is no need for a temporary deleted item, as the item
is immediately available. However, when checking if the item "exists" we
did /not/ check isDeleted, and the add_with_meta operation is rejected
with InvalidCas.

Change-Id: I4aaf18db306c8830ccea7696565e378dea5f69bb
Reviewed-on: http://review.couchbase.org/79784
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>