Improve cbstats tasks sorting by column name
Previously to sort the data by a column we needed either the column
index or the exact (but case-insensitive) name of the column.

Now we are a little bit smarter, and try to find the column with the
"closest" name.

While this is of little real benefit, it is useful for columns which
contain characters which are annoying to type at a shell (#, !, & etc.)
as they can be omitted and the closest name will be selected.

Change-Id: Ie230b8d5dd87e173c53c0ac6c8850f5c427b5d09
Reviewed-on: http://review.couchbase.org/79825
Reviewed-by: Oliver Downard <oliver.downard@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>