MB-24634: Keep buffers attached the connection
Pin the input and output buffer to the connection once it is
associated with the connection.

This avoids a buffer corruption which is caused by issues in tracking
buffer usage - conn_return_single_buffer() if curr == buff and the
number of bytes in the buffer is 0, but when we add the packet header
to the buffer we don't update curr or bytes in the buffer (we add stuff
to the buffer a number of other places as well without updating rbytes).

In master the underlying bug has been addressed by moving to a more
explicit buffer management implementaiton; but for 4.6.x we are taking
a simpler/more conservative approach and disablging this optimization.

This does have the potential to increase the memory footprint if there
are a large number of connections.

Change-Id: I41781ab763518b2bd362513c546207febdfbb486
Reviewed-on: http://review.couchbase.org/79882
Reviewed-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>