MB-24971: Signal when stuck in conn_send_data state
During bucket deletion if a connection had a full send buffer it was
possible for it to get into a conn_send_data data due to
McbpConnection::transmit() returning SoftError.

The solution is to ensure that we raise a 'fake' write event in
signalIfIdle to ensure the stalled connection has an event delivered.

The patch also modifies the testapp_bucket test
MB19748TestDeleteWhileConnShipLogAndFullWriteBuffer to
1) Increase the probablity that the test has to cope with SoftErrors
(This is achieved increasing the size of the messages being sent).
2) Ensure that if bucket deletion hangs the test terminates by
explictly closing the connection.

Change-Id: Ic214e3f293623afc9de83d9c7ec85b19cdecd745
Reviewed-on: http://review.couchbase.org/79899
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>