vattr.exptime should be 0 for no expiry, last_modified not to be negative
If a document has 0 for the expiry time, then it never expires.
The vattr code was always returning 0 + now() so it appears as if the
document had an expiry time of now().

For large CAS values the last_modified value was coming out as a
negative number, this is because cJSON uses a double internally
meaning overflowed seconds.count() was shown as a negative number.
last_modified is now a string to avoid such issues.

Change-Id: I3bea2cac56fd65771b205041691c35c6d820b373
Reviewed-on: http://review.couchbase.org/79955
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>