MB-25031: Derive max value size to defragment from allocator
Previously we used a hard coded value to determine the largest size of
document the defragmenter should relocate.

This value is that of the largest allocation for which jemalloc will
provide a region rather than whole pages. Rather than having to check if
the hard coded value needs updating each time we update jemalloc, we can
determine the size of the largest bin through the allocator hooks.

Change-Id: Ia9b53e5f98bc6bfc5439efa086faed6b93836886
Reviewed-on: http://review.couchbase.org/80011
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>