MB-25033: store_if now triggers a bgFetch for FE buckets
The store_if implementation didn't properly consider FE buckets.
In the case where there is no existing stored-value in the cache,
we are just returning success when *there* could be xattrs on disk
to merge into the new document.

We should be hitting the needs bgFetch code when we have a predicate
and nothing in cache (unless the bloom filters know better).

Change-Id: Ia5ee63a061229301017d16f8647eadb18955053d
Reviewed-on: http://review.couchbase.org/80015
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>