MB-22413: Stop skipping MB19748 Test
The test MB19748TestDeleteWhileConnShipLogAndFullWriteBuffer used to
frequently hang, and therefore it was changed to be skipped if the
TESTAPP_SKIP_HANGING_TEST environment variable was set.
See Change-Id: I331fdc6d5a10b378b0956d8499865081a7f8ca49

It is believed that the test hung due to being stuck in a
conn_send_data state.  This bug was resolved in the following patch
Change-Id: Ic214e3f293623afc9de83d9c7ec85b19cdecd745.

Therefore the test needs to no longer be skipped when the
TESTAPP_SKIP_HANGING_TEST environment variable is set.

Change-Id: I76fa26d90940b7917b6b9509f19471f4e6b3fa9b
Reviewed-on: http://review.couchbase.org/80064
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>