MB-25011: store an xattr-dirty flag in the VBucket
Store to the vbucket metadata document a boolean that indicates if
the vbucket has xattrs stored in it. This boolean is then accesible
via the VBucket object.

Note once the boolean is stored as true, no effort is made to change
to false if all of the xattr documents were fully deleted or rollback
occurred.

Change-Id: If12279265cb2423ebabfb05cc56ac800e3874cee
Reviewed-on: http://review.couchbase.org/80199
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>