MB-25011: Enhance store_if so vbucket can be queried
Change the store_if predicate signature from:

    bool pred(const item_info&);

... to:

    cb::StoreInfoStatus pred(boost::optional<item_info>&,
                             cb::vbucket_info);

So now the predicate:

1) Takes a boost::optional item_info
2) Takes a new cb::vbucket_info object
3) Can return one of 3 status codes
   i) Continue
  ii) Fail
 iii) GetInfo

In-terms of MB-25011 we really need a way to prevent store_if
doing a needless bg-fetch in full-eviction mode, that is what the
isXattrDirty() flag on the vbucket assists with.

This patch now exposes the VB::isXattrDirty bool to the store_if
predicate. With that vbucket state exposed to the predicate, it can
now decide how the store should proceed and in the case where the user
is genuinely not using xattrs, the store_if will not trigger the
read/modify/write xattr path.  The store_if predicate is also now
updated so in the case that the VB says it stores xattrs, the
predicate will request the 'target' item_info by using the GetInfo
return code, then it will only enter the read/modify/write xattr path
if the item has xattrs.

With this patch we should see a performance improvement for write
heavy workloads against full-eviction buckets when the dataset
contains no  xattrs. Testing used the following configuration.

5million 80bytes keys on a 512MB bucket.

Intial load (not measured)
> cbc-pillowfight <opts> -I 5000000 -m 80 -M 80  --populate-only

Measured phase, re-write 1million keys.
> cbc-pillowfight <opts> -I 5000000 -m 80 -M 80  --populate-only

Without this patch the set/s achieved in the measured phase was
approx 5k/s, with this patch approx 40k/s.

Change-Id: I1691b218f0c895fccd619982f965de5836497d06
Reviewed-on: http://review.couchbase.org/80200
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>