MB-25105: Ephemeral: wake up expiryPager task when high watermark reached
When an Ephemeral bucket with no-eviction reaches the high watermark,
it does not trigger the deletion of items whose TTL has expired. As a
result, memory usage (and item count) will remain at an elevated level
until the periodic expiryPager next runs (default cadence - 1hour).

The issue is that Ephemeral, no-eviction buckets do not schedule an
itemPager task - which is the task responsible for evicting
not-recently-used items from RAM. They don't schedule this task as
no-eviction buckets don't automatically delete anything - in fact
that's the main difference between these two modes - if the itemPager
task is scheduled or not.

Unfortunately it's the itemPager which /also/ handles eviction when we
hit the high watermark, hence no task runs to delete expired items.
Solution is to explicitly run the expiryPager task for ephemeral when
we hit the high watermark.

Update ItemPager tests so we run the correct task based on the bucket
config.

Change-Id: Ia660a038bad9e8df88ef3da4bb9b031054170622
Reviewed-on: http://review.couchbase.org/80241
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>