MB-25104: edit_docinfo_hook copy with the correct metadata version
The edit_docinfo_hook function is allocating V1 metadata
but then trying to copy into it using the larger V2 size, triggering
an exception from copyToBuf.

The fix is to use the V1 size, testing is added to cover this code
path (new test would throw an exception without the fix).

Change-Id: Id79e4a35a9175809a4ab1167a87f759a0afbe66d
Reviewed-on: http://review.couchbase.org/80247
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>