MB-22960 / MB-25056: Set pendingBackfill to true when backfill required
Due to cursor dropping we re-register the replication cursor only
during backfill when we mark the disk snapshot.  However by this point
it is possible that the CheckpointManager no longer contains the next
sequence number the replication stream requires (i.e. next one after
the backfill seqnos).

To avoid this data loss when we register the cursor we check to see if
the result is greater than the lastProcessedSeqno + 1.  If so we know
we may have missed some items and may need to perform another backfill.

We actually only need to do another backfill if the result is greater
than the lastProcessedSeqno + 1 and registerCursorBySeqno returns true,
indicating that the resulting seqno starts with the first item on a
checkpoint.

The patch set pendingBackfill as required, and enables / provides the
appropriate regression tests.

Change-Id: I488ec129dded11e0c1732198eab75725d8ce5b09
Reviewed-on: http://review.couchbase.org/80256
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>