MB-25105: Move EphTombstoneHTCleaner construction to EphemeralBucket ctor
Minor refactor to move the construction of EphTombstoneHTCleaner task
to EphemeralBucket ctor. This will allow tests to have access to this
task without having to call initialize (which performs scheduling
which tests may want to skip).

The extra EphemeralBucket argument to EphTombstoneHTCleaner
ctor is required because we are how constructing EphTombstoneHTCleaner
earlier, before engine->getKVBucket() has a valid pointer to return;
hence we cannot rely on that function to obtain the underlying bucket.

The test changes are necessary because we no longer have a constant
number (1) of initial nonIO tasks in all configurations - we how have
EphTombstoneHTCleaner for ephemeral.

Change-Id: Ib759451a5a49c3185b915893e1d42f95d73a64d9
Reviewed-on: http://review.couchbase.org/80260
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>