MB-25144: Check for snapStart correctly in test "dcp cursor dropping"
While streaming for items in the test "dcp cursor dropping" a check
was introduced to ensure that the 'snapStart' does not drop below
the last streamed seqno. However, there is no requirement that it must
be the immediate next seqno (next to the last streamed seqno), it can
be higher.

This commit corrects the check for 'snapStart' seqno.

Change-Id: I0ff6ed5dcf78c66e0852d77077f3c42ed2187a3e
Reviewed-on: http://review.couchbase.org/80359
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>