MB-25060: Remove bias of 1 from getNumOpenChkItems
This bias was originally added as part of MB-7435 (7923844f1), for
reasons which are not 100% clear now, but seem to be related to trying
to account for the checkpoint_start meta item in each Checkpoint -
historically such meta_items were not explicitly tracked in the
Checkpoints.

This bias, however causes problems with rebalance statistics, as it is
used for the DCP vbtakeover estimate and in that instance ns_server is
expecting an estimate of the number of non-meta items.

Fortunately we now have an accurate measure of the meta item count
(see MB-20852 / 941c24b6f), and as such we can simply return the
actual non-meta count in a Checkpoint accurately. This patch therefore
removes the bias, updating tests as appropriately.

Change-Id: I988119c61b0e7bed3670dd395d8750cf5588c6cc
Reviewed-on: http://review.couchbase.org/80370
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Sriram Ganesan <sriram@couchbase.com>