MB-24533: Fix possible hangups due to missed notifications
The testHarness.waitfor_cookie functions, waits for io notifications to
come in. But sometimes due to race conditions, the test code could end
up calling this function after the notification arrives. This was
initially fixed for MB-23990. This fix will make sure all the places
where the wait is called would be safe.

Change-Id: I20d334eb0c96bbd7dad1cbb2ae68d82937ff077c
Reviewed-on: http://review.couchbase.org/80403
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>