MB-25097: In 'Ephemeral + fail_new_data' fail rebalance on full mem usage.
In Ephemeral buckets, with fail_new_data policy, if replication or
rebalance results in a condition wherein there is not enough memory for
data items on the consumer node, then we must fail such a rebalance
or replication rather than hang.

This commit disconnects the DCP consumer connection if items
are received on it with the memory usage beyond replication
throttle (99%).

Change-Id: I0bcdb64fe254dfc989a13c7a35e9b176be8c0746
Reviewed-on: http://review.couchbase.org/80209
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>