MB-23797: Revert "Introduce testing exception for rescheduled dead tasks"
In http://review.couchbase.org/#/c/76394/ the ability to reschedule a
cancelled (and now in state TASK_DEAD) GlobalTask was fixed.

It did not appear that any tasks other than the ItemPager for
ephemeral buckets are rescheduled once dead, but to avoid changing
existing behaviour a "paranoia" check was introduced (6e6429b) to
throw an exception if any task other than the ItemPager is
rescheduled, to bring them to our attention.

This patch has had 3 months of soak time, and hasn't exposed any
unexpected wakups, but to ensure we don't fire such exceptions in
production the check is being reverted.

This reverts commit 6e6429b0b1e20aa44211452198a7b5c80b9ae835.

Change-Id: I82f8d4068e03c451493ef7ac17bd3d21c37f4166
Reviewed-on: http://review.couchbase.org/80594
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>