MB-20940: [9/n] Tap Removal - Remove TapConnMap Class
As part of Tap Removal, this patch removes the TapConnMap class.
In addition it removes classes that are only referenced from
TapConnMap class.

The use of TapConnMap is also removed from EventuallyPersistentEngine.

The functions processAck, setTimeForNoop and registerCursor are also
removed from the TapProducer class as they reference TapConnMap.  The
setBackfillPhase function from the Consumer class and
processCheckpointCommand from the TapConsumer class, which calls
setBackfillPhase, are also removed.

With the removal of tapconnmap.h, the kv_bucket.cc includes ep_engine.h
so it can access EventuallyPersistentEngine.

Change-Id: Icd61776b53a9e141bcc38ea628a845dc939cd65b
Reviewed-on: http://review.couchbase.org/80990
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>