MB-25273: Fix dcp_delete with xattrs
The deleteWithMeta function is shared between DCP delete and mcbp
del_with_meta. The deleteWithMeta function has xattr pruning support
intended for mcbp del_with_meta, however this code was triggering
against the DCP call, triggering the exception seen in the MB where
a new mutation is pushed with a seqno below the checkpoint.

There are two fixes applied here for this exception.

1) deleteWithMeta skips pruning of xattrs for DCP callers, however
   2 below now means the DCP delete with body case doesn't now go down
   this path.
2) The consumer DCP deletion handler now pushes a delete with value
   down the mutation path, similar to how the front-end handles a
   delete with value.

Finally after the exception is fixed it was noted that a replication
mismatch occurs. When the active applies the delete and keeps the
system xattrs in a deleted document, the flags of the original item
were preserved. As the delete is sent to the replica using DCP_DELETE
which does not replicate flags, the replica document would be near
identical, except for having flags = 0.

Now as memcached creates the delete with value (preserving system
xattrs) it forces the flags to zero on the delete with value item.
This should be acceptable as the only code that can actually read
back the deleted document is the SUBDOC with the new "access-deleted"
flag, so shouldn't be surprising behavior for legacy users.

Change-Id: I5ba49ffb3bfff2685a1ea9d606127ce6636204ea
Reviewed-on: http://review.couchbase.org/81165
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>