MB-20940: [17/n] Tap Removal - Move TapParam to ReplicationParam
The Tap Parameters consisted of the following:
- tap_keepalive
- replication_throttle_queue_cap
- replication_throttle_cap_pcnt
- replication_throttle_threshold

With the removal of Tap the tap_keepalive parameter is no longer
required and has been removed.

The remaining three parameters are still required and therefore
have been renamed as replication parameters.

The mc_bin_client has been updated to reflect the rename.

The now unused MAX_TAP_KEEP_ALIVE constant is removed.

Change-Id: I26b637c4a5d8241a7b20018534b41bf14ffcf250
Reviewed-on: http://review.couchbase.org/81106
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>