MB-20940: [18/n] Tap Removal - Remove the Producer class
As part of Tap Removal, this patch removes the Producer class.  Most
of the functions are removed, however the small number that are
required are moved to the DcpProducer class.

Also removed the reconnects stat which is only used by TAP.
The ep_dcp_queue_backfillremaining stat is also never incremented
in DCP and so is removed.

Change-Id: Ib32e9caa0c376ee2ab954d4410fa6d8217e6fe48
Reviewed-on: http://review.couchbase.org/81166
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>