MB-25179: Fix write buffer management
There was a bug in handling of the write buffers when trying
to determinate if the buffer was still in use or not. It would
manifest itself under high write load to the socket from the
memcached side.

Each connection would use the write buffer object as a scratch
space where it could format data into, and then record the
offset into the buffer into the io-vector for data to send
over the network by using sendmsg. So far everything sounds
good, but at one point we wanted to reduce the resource usage
from idle connections by "stealing" the read and write buffer
from the connection if there wasn't any pending data in the
buffer. For the read buffer this check was correct, but
we had code paths in the write buffer path which wouldn't
update the number of bytes it had stored in the buffer.
Under a "normal" load this wouldn't fail, as we normally only
used 24ish bytes in the buffer (the protocol header), and
the socket would accept this chunk of memory (but possibly
not the large value following located in _another_ buffer
pinned to the request and not the write buffer).

With this patch we track how much data we format into the buffer,
and how much data we read out of the buffer so that we know for
sure when we're done with the buffer.

Change-Id: I39c4175034511812099a024d178b5195450c5538
Reviewed-on: http://review.couchbase.org/81177
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>