MB-25097: In 'Ephemeral + fail_new_data' fail rebalance on hitting ENOMEM
In Ephemeral buckets, with fail_new_data policy, if replication or
rebalance results in a condition wherein there is not enough memory for
data items on the consumer node, then we must fail such a rebalance
or replication rather than hang.

We cannot always easily estimate the new memory used in the initial
stages of processing a replicated item. When we hit full memory
(ENOMEM) just before adding item to the vbucket, the status is
returned to the connection level. The connection then should
handle the ENOMEM status accordingly, that is, disconnect in case of
'Ephemeral + fail_new_data' policy.

Change-Id: I1aea12fa16dd1b076b79ed1e70dc8e62339b6f72
Reviewed-on: http://review.couchbase.org/81217
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>