MB-22387: Add a separate return type for add temporary items
The addTempStoredValue method only returns out of memory (NoMem)
and background fetch needed (BgFetch). Add a new return status
for adding temporary items instead of using AddStatus which
returns 6 different values and therefore 6 cases need to be
handled on return. Also, make the return type an enum, so that
it is extensible.

Change-Id: I4717fd405b00663f71eca7b1d665ed2152b6e4b9
Reviewed-on: http://review.couchbase.org/81314
Reviewed-by: Daniel Owen <owend@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>