MB-20940: [23/n] Tap Removal - Clean-up ConnHandler Class
Removes the following unused variables:
- supportCheckpointSync_
- connToken
- expiryTime
- counter_

Removes the following unused methods:
- engine()
- supportsAck()
- setSupportCheckpointSync(bool checkpointSync)
- supportsCheckpointSync()
- processedEvent(uint16_t event, ENGINE_ERROR_CODE ret)
- setName(const std::string &n)
- setCookie(const void *c)
- setExpiryTime(rel_time_t t)
- getExpiryTime()
- getLastWalkTime()
- setConnected(bool s)
- getAnonName()
- getConnectionToken()

setLastWalkTime() is used but the lastWalkTime variable is never read.
Therefore both are removed.

isConnected() is used but the connected variable is only ever set to
true at construction and therefore is of no value and hence is removed
as is the associated 'stat'.

releaseReference takes an optional bool that is never used, and
therefore is removed.

numDisconnects is always zero and therefore is removed.

Change-Id: I0e6e483b8de67c61a26794c6cf4e74d50fdf4a89
Reviewed-on: http://review.couchbase.org/81344
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>