MB-25398: Allow store_deleted_with_CAS on non-resident items
When attempting to perform a store operation with a deleted body to a
deleted key, with CAS (!), allow it to succeed if the key is no longer
resident. This is the case if the flusher has completed storing the
delete to disk; and hence the persistence callback has run and removed
the deleted item from the HashTable.

Note this manifested as only an occasional failure when trying to
perform:

    1. CAS = delete(key)
    2. subdoc_multi_mutate(key, AccessDeleted=True, new_body=...,
                           in_CAS=cas)

If 1 and 2 executed back-to-back without the flusher running in
between, then the request succeeded, as the deleted item from (1)
would still be resident in the HashTable when we execute (2). However,
if the flusher ran between those two statements, then (2) would fail
with ENOENT.

Change-Id: I4a7e213469a434d437bcab5c78ff34a189ccb5ca
Reviewed-on: http://review.couchbase.org/81479
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Manu Dhundi <manu@couchbase.com>