MB-20940: [24/n] Tap Removal - Remove Notifiable Class
Removes Notifiable:notificationScheduled and the associated methods
because although the variable was set to true/false, it was never read.

Removes Notifiable:notifySent and the assosciated methods because
although the variable was set to true/false, it was never read.

Removes Notifiable:suspended and the associated methods because the
variable is never set or read.

Th remaining paused and pausedReason variables in the Notifiable class
have been moved to the ConnHandler class, allowing the paused and
pausedReason stats to be moved from DcpConsumer and DcpProducer into
a single place in the ConnHandler class.

The Notifiable class contained no other variables and hence has been
removed.

Change-Id: I52bbff55f0fece672600aab0461e9c4d955c6d74
Reviewed-on: http://review.couchbase.org/81428
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>