MB-25097: In 'fail_new_data' disconnect from processor tsk on full mem
In Ephemeral buckets, with fail_new_data policy, if replication or
rebalance results in a condition wherein there is not enough memory for
data items on the consumer node, then we must fail such a rebalance
or replication rather than hang.

It is possible that the memory usage can be found above the threshold
while we are processing the items from the consumer buffer in the
consumer processor task. Upon hitting such a condition, the processor
task stops processing and notifies the DCP consumer to disconnect.

Change-Id: If8ecbce3c50a09f621d7460307b13e59e927eb07
Reviewed-on: http://review.couchbase.org/81270
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>