MB-25097: In 'fail_new_data' disconnect from processor tsk on ENOMEM
In Ephemeral buckets, with fail_new_data policy, if replication or
rebalance results in a condition wherein there is not enough memory for
data items on the consumer node, then we must fail such a rebalance
or replication rather than hang.

It is possible that the memory usage can be found above the threshold
while we are processing the items from the consumer buffer in the
consumer processor task.

Further, we cannot always easily estimate the new memory used in the
initial stages of processing a replicated item. When we hit full memory
(ENOMEM) just before adding item to the vbucket, the status is
returned to the processor task level. Then the processor task stops
processing and notifies the DCP consumer to disconnect.

Change-Id: I8ae9c39c2cc7669d056ca0e1ee90e8e7b10885a1
Reviewed-on: http://review.couchbase.org/81374
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>