MB-25434: Move Backfilling TMPFAIL check to after hard fail checks
During a streamRequest we currently check if we are in a backfilling
state (which results in a TMPFAIL) before checks such as whether the
seqnos are in range (which results in a permanent error - ERANGE).

It is better to fail on permanent errors before failing on checks that
result in a TMPFAIL.  Therefore the backfilling state check is moved to
after the permanent error checks.

Change-Id: I92321abfaaba8f204059c8daeaa6c067acbb3f9f
Reviewed-on: http://review.couchbase.org/81599
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Oliver Downard <oliver.downard@couchbase.com>