MB-20940: [27/n] - Tap Removal - Rename tapconnection.h/.cc
Given the removal of TAP, the tapconnection files now contain the
ConnContainer and ConnHandler class.

Given the main class is ConnHandler, the files have been remamed
connhandler.h and conhandlier.cc.

In addition the affected includes have been renamed.  This required the
kv_bucket_iface.h to be added to conhandler.h, as prevously it was
getting this include indirectly from an earlier included file, however
is no longer the case.

Change-Id: I9d39e68ba0d8d6eaf1d709961f6c070fd55282fa
Reviewed-on: http://review.couchbase.org/81614
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>