MB-25342: Reorder Collections::VB::Manifest update/serialise
Previous to this commit an update to the Collections::VB::Manifest
would:

1: Serialise the manifest state to a system-event
2: Update the manifest

This ordering is creating issues for new code that is performing full
deletion of the collection data. The issue being that some new changes
being made by step 2 don't get captured in step 1.

It is much simpler to (and really should of been like this from the
start):

1: Update the manifest
2: Serialise the manifest state to a system-event

This patch refactors the add/beginDelete methods so that they return
the ManifestEntry which is affected by the update allowing us to
"patch" it with the seqno we get from queueing of the serialised
manifest.

::completeDeletion is also updated but not refactored.

VbucketManifestTest is also altered slightly so that completeDeletion
is given more test coverage, it now applies updates to the replica
and compares active with replica.

Change-Id: I7361c5675356e47888afcf751dc4db21ac03762e
Reviewed-on: http://review.couchbase.org/81983
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Will Gardner <willg@rdner.io>
Reviewed-by: Dave Rigby <daver@couchbase.com>