MB-25402: Fix conversion of expiration times
Memcached buckets returns the relative expiration time as part
of the get_item_info method. Couchbase buckets returns the
absolute expiration time. Given that the core only cares about
the expiration time in two cases (incr/decr to preserve the
expiration time, and subdoc where we return the meta information
as virtual xattrs) it'll be less intrusive to make the conversion
where we actually need it instead of doing it always and throw it
away.

Change-Id: I8be18b40ca67c933b55c9b3eb358e8eae2f99a21
Reviewed-on: http://review.couchbase.org/81841
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>